Refactor: SalaryData 타입 별도 파일로 분리
