+++ src/pages/mypage/salaryTypes.ts
@@ -1,0 +1,17 @@
+export interface SalaryData {
+  id: string;
+  date: string;
+  rawDate: number;
+  base: number;
+  bonus: number;
+  position: number;
+  overtime: number;
+  night: number;
+  health: number;
+  care: number;
+  job: number;
+  tax: number;
+  totalPayment: number;
+  totalDeduct: number;
+  actualPayment: number;
+}
