+++ src/pages/mypage/ui/salary-modal.tsx
@@ -2,7 +2,7 @@ import { useMemo } from 'react';
 import { useNavigate } from 'react-router-dom';
 import * as S from '@/pages/mypage/styles/salary-modal.styles';
 import Button from '@/shared/button/Button';
-import { SalaryData } from '@/pages/mypage/ui/salary-section';
+import { SalaryData } from '../salaryTypes';
 import { formatCurrency } from '@/utils/formatCurrency';
 import { formatDate } from '@/utils/formatDate';
 import Lottie from 'lottie-react';
